% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_io_utils.R
\name{filter_blacklist}
\alias{filter_blacklist}
\title{Filter ENCODE Blacklist Regions}
\usage{
filter_blacklist(gr, genome = "hg38", blacklist_path = NULL)
}
\arguments{
\item{gr}{A GRanges object.}

\item{genome}{Character. "hg38", "hg19", "mm10", "mm9".}

\item{blacklist_path}{Custom BED path or NULL (uses built-in).}
}
\value{
Filtered GRanges.
}
\description{
Filter ENCODE Blacklist Regions
}
\examples{
gr <- GenomicRanges::GRanges("chr1", IRanges::IRanges(c(100, 5000000), width = 200))
filter_blacklist(gr, genome = "hg38")
}
