% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_object_contract.R
\name{get_combined_class_results}
\alias{get_combined_class_results}
\title{Extract Combined Class Results}
\usage{
get_combined_class_results(se)
}
\arguments{
\item{se}{A SummarizedExperiment after \code{combine_superdomain_calls}.}
}
\value{
A \code{data.frame} of combined classes (and Combined_Domain_Class if
present).
}
\description{
Extract Combined Class Results
}
\examples{
data(example_se)
se <- call_super_domains(example_se, feature = "Intensity", verbose = FALSE)
se <- call_super_domains(se, feature = "Breadth", verbose = FALSE)
se <- combine_superdomain_calls(se)
head(get_combined_class_results(se))
}
