% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_consensus.R
\name{get_consensus_peaks}
\alias{get_consensus_peaks}
\title{Generate Consensus Peak Set for epiPortrait}
\usage{
get_consensus_peaks(gr_list, min_reps = 2)
}
\arguments{
\item{gr_list}{A GRangesList or list of GRanges objects containing peaks from all samples.}

\item{min_reps}{Minimum number of replicates a peak must appear in to be retained (default: 2).}
}
\value{
A filtered GRanges object representing the consensus peaks.
}
\description{
Merges multiple peak sets into a unified genomic coordinate system
while applying a reproducibility filter.
}
