% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_object_contract.R
\name{get_sample_results}
\alias{get_sample_results}
\title{Extract Sample-Level Results / QC Table}
\usage{
get_sample_results(se)
}
\arguments{
\item{se}{A SummarizedExperiment.}
}
\value{
A \code{data.frame} with one row per sample (colData + QC).
}
\description{
Extract Sample-Level Results / QC Table
}
\examples{
data(example_se)
get_sample_results(example_se)
}
