% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_object_contract.R
\name{get_transition_results}
\alias{get_transition_results}
\title{Extract Transition Results}
\usage{
get_transition_results(se)
}
\arguments{
\item{se}{A SummarizedExperiment after \code{compare_superdomains} /
\code{compare_superdomain_classes}.}
}
\value{
A \code{data.frame} of transition columns.
}
\description{
Extract Transition Results
}
\examples{
data(example_se)
se <- call_super_domains(example_se, feature = "Intensity",
                         mode = "per_group", group_var = "Condition",
                         verbose = FALSE)
se <- compare_superdomains(se, group_var = "Condition",
                           ref_group = "Control", target_group = "Treatment")
head(get_transition_results(se))
}
