% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_object_contract.R
\name{summarize_epiportrait_object}
\alias{summarize_epiportrait_object}
\title{Summarize an epiPortrait Object}
\usage{
summarize_epiportrait_object(se)
}
\arguments{
\item{se}{A SummarizedExperiment.}
}
\value{
A list with a short text summary of the object contents.
}
\description{
Summarize an epiPortrait Object
}
\examples{
data(example_se)
summarize_epiportrait_object(example_se)
}
