% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_utils.R
\name{filter_promoter_peaks}
\alias{filter_promoter_peaks}
\title{Filter Peaks by Genomic Annotations (Promoter Exclusion)}
\usage{
filter_promoter_peaks(gr, genome = "hg38", upstream = 2000, downstream = 2000)
}
\arguments{
\item{gr}{A GRanges object of peaks.}

\item{genome}{A character string ("hg38", "hg19", "mm10", etc.) or a TxDb object.}

\item{upstream}{Number of bp upstream of TSS to define promoter (default: 2000).}

\item{downstream}{Number of bp downstream of TSS to define promoter (default: 2000).}
}
\value{
A filtered GRanges object.
}
\description{
Filters out peaks that overlap with specified genomic regions,
typically used to remove promoter-proximal peaks before Super-Element analysis.
}
