% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_utils.R
\name{stitch_epi_peaks}
\alias{stitch_epi_peaks}
\title{Stitch Proximal Epigenetic Peaks}
\usage{
stitch_epi_peaks(gr, stitch_distance = 12500)
}
\arguments{
\item{gr}{A GRanges object of peaks (typically the output from get_consensus_peaks).}

\item{stitch_distance}{Numeric. Maximum distance (in base pairs) between peaks
to be stitched together. For Super-Elements, 12500 (12.5 kb) is standard.
For standard broad marks, you might use 3000 to 5000.}
}
\value{
A GRanges object of stitched domains, with a metadata column
'Constituent_Peaks' recording how many original peaks were merged into each domain.
}
\description{
Merges closely spaced peaks into larger continuous macro-domains based on
a user-defined stitching distance. This is crucial for analyzing Super-Elements
or broad heterochromatin spreading in epiPortrait.
}
